Completed Development according to course playlist

+++ frontend/src/components/WorkoutDetails.js
@@ -1,10 +1,19 @@
 import { UseWorkoutsContext } from "../hooks/UseWorkoutsContext";
 import { formatDistanceToNow } from "date-fns";
+import { UseAuthContext } from "../hooks/UseAuthContext";
+
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = UseWorkoutsContext();
+  const { user } = UseAuthContext();
   const DeleteOnClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.createdToken}`,
+      },
     });
     const jsonForm = await response.json();
 
@@ -25,7 +34,7 @@ const WorkoutDetails = ({ workout }) => {
         {formatDistanceToNow(workout.createdAt, { addSuffix: true })}
         <br />
       </p>
-      <span class="material-symbols-outlined" onClick={DeleteOnClick}>
+      <span className="material-symbols-outlined" onClick={DeleteOnClick}>
         delete
       </span>
     </div>
